Rename cookie store implements #2
