+++ http/cookie/file.go
@@ -1,4 +1,4 @@
-package cookiestore
+package cookie
 
 import (
 	"encoding/gob"
@@ -7,7 +7,7 @@ import (
 	"os"
 )
 
-func Load(url *url.URL, file string, jar http.CookieJar) error {
+func LoadFrom(url *url.URL, file string, jar http.CookieJar) error {
 	f, err := os.Open(file)
 	if err != nil {
 		return nil
@@ -21,7 +21,7 @@ func Load(url *url.URL, file string, jar http.CookieJar) error {
 	return nil
 }
 
-func Save(url *url.URL, file string, jar http.CookieJar) error {
+func SaveTo(url *url.URL, file string, jar http.CookieJar) error {
 	cookies := jar.Cookies(url)
 	if len(cookies) == 0 {
 		return nil
